requirements.txt
playwright
beautifulsoup4
openpyxl
python-docx
requests
streamlit